+++ test_igseq/test_trim.py
@@ -1,9 +1,94 @@
 from tempfile import TemporaryDirectory
 from pathlib import Path
-from igseq.trim import trim
+from igseq.trim import trim, get_adapters_fwd, get_adapter_rev
+from igseq.util import IgSeqError
 from .util import TestBase, TestLive
 
-class TestTrimLive(TestBase, TestLive):
+class TestTrimHelper(TestBase):
+    """Helper functions for the actual trimming-related tests"""
+
+    def check_trim_output(self, dir_obs, dir_exp):
+        """Check that an observed trim output dir matches an expected dir"""
+        dir_obs = Path(dir_obs)
+        dir_exp = Path(dir_exp)
+        files = sorted([p.name for p in dir_obs.glob("*")])
+        files_expected = sorted([p.name for p in dir_exp.glob("*")])
+        self.assertEqual(files_expected, files)
+        for path in files_expected:
+            if path.endswith(".gz"):
+                self.assertGzipsMatch(dir_obs/path, dir_exp/path)
+            if path.endswith(".counts.csv"):
+                self.assertTxtsMatch(dir_obs/path, dir_exp/path)
+
+class TestTrim(TestTrimHelper):
+    """Basic tests of trim-related functions"""
+
+    def setUp(self):
+        super().setUp()
+        self.sample = {
+            "Sample": "sample1",
+            "BarcodeFwdSeq": "NNNNAACCACTA",
+            "BarcodeRevSeq": "TAGTGGTT",
+            "Type": "gamma"}
+
+    def test_get_adapters_fwd(self):
+        """Test getting the adapter sequences to trim from the end of R1"""
+        with self.subTest(case="basic"):
+            adapter_fwd = get_adapters_fwd(self.sample, "rhesus")
+            self.assertEqual(
+                adapter_fwd,
+                {"rhesus_gamma": "TCCACCAAGGGCCCATCGGTCTTCCCCCTGGC"})
+        with self.subTest(case="bad species"):
+            with self.assertRaises(IgSeqError):
+                adapter_fwd = get_adapters_fwd(self.sample, "rhino")
+        with self.subTest(case="bad chain type"):
+            samp = self.sample.copy()
+            samp["Type"] = "heavy"
+            with self.assertRaises(IgSeqError):
+                adapter_fwd = get_adapters_fwd(samp)
+        with self.subTest(case="no species"):
+            adapter_fwd = get_adapters_fwd(self.sample)
+            self.assertEqual(
+                adapter_fwd, {
+                "rhesus_gamma": "TCCACCAAGGGCCCATCGGTCTTCCCCCTGGC",
+                "human_gamma": "TCCACCAAGGGCCCATCGGTCTTCCCCCTGGC"})
+        with self.subTest(case="no chain type"):
+            adapter_fwd = get_adapters_fwd({}, "rhesus")
+            self.assertEqual(
+                adapter_fwd, {
+                "rhesus_gamma": "TCCACCAAGGGCCCATCGGTCTTCCCCCTGGC",
+                "rhesus_alpha": "CCGACCAGCCCCAAGGTCTTCCCGCTGAGCCTC",
+                "rhesus_mu": "TGCATCCGCCCCAACCCTTTTCCCCCTCGTCTC",
+                "rhesus_epsilon": "CACACAGAGCCCATCCGTCTTCCCCTTGACCCG",
+                "rhesus_delta": "CCAAGGCTCCGGATGTGTTCCCCATCATATCAG",
+                "rhesus_kappa": "CTGTGGCTGCACCATCTGTCTTCATCTTCCCGC",
+                "rhesus_lambda": "CCCTCCTCTGAGGAGCTTCAAGCCAACAAGGCC"})
+        with self.subTest(case="blank chain type"):
+            adapter_fwd = get_adapters_fwd({"Type": ""}, "rhesus")
+            self.assertEqual(
+                adapter_fwd, {
+                "rhesus_gamma": "TCCACCAAGGGCCCATCGGTCTTCCCCCTGGC",
+                "rhesus_alpha": "CCGACCAGCCCCAAGGTCTTCCCGCTGAGCCTC",
+                "rhesus_mu": "TGCATCCGCCCCAACCCTTTTCCCCCTCGTCTC",
+                "rhesus_epsilon": "CACACAGAGCCCATCCGTCTTCCCCTTGACCCG",
+                "rhesus_delta": "CCAAGGCTCCGGATGTGTTCCCCATCATATCAG",
+                "rhesus_kappa": "CTGTGGCTGCACCATCTGTCTTCATCTTCCCGC",
+                "rhesus_lambda": "CCCTCCTCTGAGGAGCTTCAAGCCAACAAGGCC"})
+
+    def test_get_adapter_rev(self):
+        """Test getting the adapter sequence to trim from the end of R2"""
+        adapter_rev = get_adapter_rev(self.sample)
+        self.assertEqual(adapter_rev, "TAGTGGTTNNNNAGATCGGAAGAGCGTCGTGTAGGGAAAGA")
+
+    def test_trim(self):
+        """Test the high-level trim() function"""
+        self.skipTest("not yet implemented")
+
+    def test_trim_pair(self):
+        """Test the lower-level single-sample trim_pair() function"""
+        self.skipTest("not yet implemented")
+
+class TestTrimLive(TestTrimHelper, TestLive):
     """Basic tests of trim with actual cutadapt.
 
     Here we have a simple case with two sample with perfect adapters in R1 and
@@ -24,17 +109,7 @@ class TestTrimLive(TestBase, TestLive):
         """Test that adapters are trimmed from R1 and R2 pairs with dir input."""
         with TemporaryDirectory() as temp:
             trim([self.path/"input/run"], self.path/"samples.csv", dir_out=temp)
-            files = sorted([p.name for p in Path(temp).glob("*")])
-            #import shutil
-            #for path in files:
-            #    shutil.copy(Path(temp)/path, ".")
-            files_expected = sorted([p.name for p in (self.path/"output").glob("*")])
-            self.assertEqual(files_expected, files)
-            for path in files_expected:
-                if path.endswith(".gz"):
-                    self.assertGzipsMatch(Path(temp)/path, self.path/"output"/path)
-                if path.endswith(".counts.csv"):
-                    self.assertTxtsMatch(Path(temp)/path, self.path/"output"/path)
+            self.check_trim_output(temp, self.path/"output")
 
     def test_trim_file_input(self):
         """Test that adapters are trimmed from R1 and R2 pairs with file input."""
@@ -44,12 +119,127 @@ class TestTrimLive(TestBase, TestLive):
                     self.path/f"input/run/sample{idx}.R1.fastq.gz",
                     self.path/f"input/run/sample{idx}.R2.fastq.gz"],
                     self.path/"samples.csv", dir_out=temp)
-            files = sorted([p.name for p in Path(temp).glob("*")])
-            files_expected = sorted([p.name for p in (self.path/"output").glob("*")])
-            self.assertEqual(files_expected, files)
-            for path in files_expected:
-                if path.endswith(".gz"):
-                    self.assertGzipsMatch(Path(temp)/path,
-                            self.path/"output"/path)
-                if path.endswith(".counts.csv"):
-                    self.assertTxtsMatch(Path(temp)/path, self.path/"output"/path)
+            self.check_trim_output(temp, self.path/"output")
+
+    def test_trim_custom_args(self):
+        """Test giving custom cutadapt arguments"""
+        with TemporaryDirectory() as temp:
+            temp = Path(temp)
+            trim([
+                self.path/"input/run/sample1.R1.fastq.gz",
+                self.path/"input/run/sample1.R2.fastq.gz"],
+                self.path/"samples.csv", dir_out=temp,
+                extra_cutadapt_args=["--too-short-output", temp/"short.fastq.gz"])
+            # there actually should be no too-short sequences, but, the file
+            # should still be created (even though empty)
+            self.assertEmpty(temp/"short.fastq.gz")
+
+    def test_trim_custom_adapters(self):
+        """Test using custom adapters for trimming"""
+        with self.subTest(case="basic"):
+            # give the actual forward and reverse adapters appropriate for each
+            # sample (that should be what would have been used anyway) and
+            # check that the output is the same as we get from the other
+            # test_trim_... cases above
+            forwards = [
+                "^AAGCAGTGGTATCAACGCAGAGTACATGGG...TCCACCAAGGGCCCATCGGTCTTCCCCCTGGC",
+                "^AAGCAGTGGTATCAACGCAGAGTACATGGG...CACACAGAGCCCATCCGTCTTCCCCTTGACCCG",
+                "^AAGCAGTGGTATCAACGCAGAGTACATGGG...TCCACCAAGGGCCCATCGGTCTTCCCCCTGGC",
+                "^AAGCAGTGGTATCAACGCAGAGTACATGGG...TCCACCAAGGGCCCATCGGTCTTCCCCCTGGC"]
+            reverses = [
+                "TAGTGGTTNNNNAGATCGGAAGAGCGTCGTGTAGGGAAAGA",
+                "TTAGAGTTNNNNNAGATCGGAAGAGCGTCGTGTAGGGAAAGA",
+                "AGGGCCTTNNNNNNAGATCGGAAGAGCGTCGTGTAGGGAAAGA",
+                "GACATATTNNNNNNNAGATCGGAAGAGCGTCGTGTAGGGAAAGA"]
+            with TemporaryDirectory() as temp:
+                temp = Path(temp)
+                for idx in [1, 2, 3, 4]:
+                    trim([
+                        self.path/f"input/run/sample{idx}.R1.fastq.gz",
+                        self.path/f"input/run/sample{idx}.R2.fastq.gz"],
+                        self.path/"samples.csv", dir_out=temp,
+                        custom_adapter_fwd=forwards[idx-1],
+                        custom_adapter_rev=reverses[idx-1])
+                self.check_trim_output(temp, self.path/"output")
+        with self.subTest(case="custom"):
+            # using custom adapters for both forward and reverse, where the
+            # output should be different for all samples now.  Untrimmed reads
+            # will not be automatically filtered since the forward primer isn't
+            # anchored
+            with TemporaryDirectory() as temp:
+                temp = Path(temp)
+                for idx in [1, 2, 3, 4]:
+                    trim([
+                        self.path/f"input/run/sample{idx}.R1.fastq.gz",
+                        self.path/f"input/run/sample{idx}.R2.fastq.gz"],
+                        self.path/"samples.csv", dir_out=temp,
+                        # an arbitrary chunk near the end of sample 1 and 3's
+                        # R1s
+                        custom_adapter_fwd="GTGAAGGTTTACCGGCATTCAT",
+                        # a partial chunk of the P5 universal seq
+                        custom_adapter_rev="GTCGTGTAGGGAAAGA")
+                    self.assertNotInFile(temp/f"sample{idx}.cutadapt2.json", "--discard-untrimmed")
+                self.check_trim_output(temp, self.path/"output-custom-adapters")
+        with self.subTest(case="custom anchored"):
+            # using custom adapters for both forward and reverse, with the
+            # forward primer linked with the first part anchored to the start
+            # of the sequence.  In this case it should be interpreted as
+            # required, and untrimmed reads filtered out.
+            with TemporaryDirectory() as temp:
+                temp = Path(temp)
+                for idx in [1, 2, 3, 4]:
+                    trim([
+                        self.path/f"input/run/sample{idx}.R1.fastq.gz",
+                        self.path/f"input/run/sample{idx}.R2.fastq.gz"],
+                        self.path/"samples.csv", dir_out=temp,
+                        # anchored, but different from the default; includes a
+                        # bit of read1 and read4 so only these should be kept
+                        # (plus 20 random NT at the 3' adapter end that should
+                        # have no effect)
+                        custom_adapter_fwd=("^AAGCAGTGGTATCAACGCAGAGTACATGGGGAGGCTC"
+                            "...GACGGCCCTTTTCAAACCAA"),
+                        # a partial chunk of the P5 universal seq (same as the
+                        # previous case)
+                        custom_adapter_rev="GTCGTGTAGGGAAAGA")
+                    self.assertInFile(temp/f"sample{idx}.cutadapt2.json", "--discard-untrimmed")
+                self.check_trim_output(temp, self.path/"output-custom-adapters-anchored")
+
+    def test_trim_edge_case(self):
+        """Test an indel-related edge case that depends on particular cutadapt behavior"""
+        # This is an example of a read pair whose trimming output changes as of
+        # cutadapt 4.0.
+        #
+        # https://cutadapt.readthedocs.io/en/stable/changes.html#v4-0-2022-04-13
+        # https://cutadapt.readthedocs.io/en/stable/algorithms.html#algorithm-indel-scores
+        with TemporaryDirectory() as temp:
+            trim([
+                self.path/"input-edgecase/run/sample.R1.fastq.gz",
+                self.path/"input-edgecase/run/sample.R2.fastq.gz"],
+                self.path/"samples.edgecase.csv", dir_out=temp)
+            self.check_trim_output(temp, self.path/"output-edgecase")
+
+class TestTrimLiveNoChainType(TestTrimHelper, TestLive):
+    """Test trimming without specifying chain type (for the constant region primer).
+
+    In this case cutadapt should fall back on using all applicable primer
+    sequences (per-species, if specified, or across species, if not).  The test
+    files are the same as TestTrimLive, where sample 4 R1 happens to end with
+    CCG which, here, ends up matching the alpha primer and getting slightly
+    trimmed.
+    """
+
+    def test_trim_dir_input(self):
+        """Test that adapters are trimmed from R1 and R2 pairs with dir input."""
+        with TemporaryDirectory() as temp:
+            trim([self.path/"input/run"], self.path/"samples.csv", dir_out=temp)
+            self.check_trim_output(temp, self.path/"output")
+
+    def test_trim_file_input(self):
+        """Test that adapters are trimmed from R1 and R2 pairs with file input."""
+        with TemporaryDirectory() as temp:
+            for idx in [1, 2, 3, 4]:
+                trim([
+                    self.path/f"input/run/sample{idx}.R1.fastq.gz",
+                    self.path/f"input/run/sample{idx}.R2.fastq.gz"],
+                    self.path/"samples.csv", dir_out=temp)
+            self.check_trim_output(temp, self.path/"output")